Ensure that all of the float/double/decimal distributions are about the same.
Scratch, when run, shows that the fraction difference between the largest bucket of 256 and the smallest is always fairly small; the highest difference is just over 0.1 (for NextExclusiveDouble()) and all the others are very close to 0.8, sometimes reaching near 0.7 or 0.9 for some seeds. BasicTest is adjusted for this and accommodates fractions up to 0.11, so the largest bucket is no more than 1/9 larger than the smallest.