Use "Bitsy" method for exclusive FP values.
This should be a little faster on the supported platforms, and more solid in terms of my understanding of how it works. The old way was pretty cool that it worked at all, though!

This commit also drops support for .NET 5.0, .NET Standard 2.1, and .NET Core 3.1 . I don't actually know what Core 3.1 is, but it is recent enough that I could support it if necessary. I'd rather not, because some APIs I want to use moving forward aren't supported by Core 3.1 .